feat(org-details): Refactoring components;
This patch adds two new components called `<Logo/>` and `<Tags/>` under
`<Directory/>`. The idea here is that these would be reusable Tailwind styles
within the design concept of a Organization. I may propose we rename some of
these components once I have everything in place. There's also an unimplemented
component called `<Header/>` which will be used for the heading text for the
Organization, but I'm not sure and may remove it in later commits.