feat: Improving how internal links are passed;
By hoisting up the `internalLinks`, we can pass them as props to the necessary
components since we have two components now for the navigation. _I do think we
should revisit this_ and possibly remove ShadCDN, but for now we have something
that is working before the launch. [See the ticket for more info][1] 🔒.

[1]: https://app.asana.com/0/1207992233793749/1208694960699967