fix: Updating the colors of the logo & logoline;
These were not appearing properly across light and dark modes. Because the
mobile menu and the non-mobile menu are different, I've added props to modify
things. I've also brought in an SVG component for the Logo and am keeping it in
the `<FullBrand/>` component for now. This kind of SVG is much better than
working with Next's `<Image/>` component. Next also doesn't optimize SVGs at
all.